Move mock PLOT data out of checkout view model
