+++ Library/ViewModels/NoShippingPledgeViewModel.swift
@@ -1285,20 +1285,3 @@ private func pledgeAmountSummaryViewData(
     rewardIsLocalPickup: rewardIsLocalPickup
   )
 }
-
-// TODO: Remove this when implementing the API [MBL-1838](https://kickstarter.atlassian.net/browse/MBL-1838)
-public func mockPledgePaymentIncrement() -> [PledgePaymentIncrement] {
-  var increments: [PledgePaymentIncrement] = []
-  #if DEBUG
-    var timeStamp = TimeInterval(1_733_931_903)
-    for _ in 1...4 {
-      timeStamp += 30 * 24 * 60 * 60
-      increments.append(PledgePaymentIncrement(
-        amount: PledgePaymentIncrementAmount(amount: 250.0, currency: "USD"),
-        scheduledCollection: timeStamp
-      ))
-    }
-  #endif
-
-  return increments
-}
